[refact] application excel 부분 분리 #97
